requirements.txt
asyncpg==0.20.1
kafka-python==2.0.1
